refactor : pages폴더 내부에서 container와 view분리 방식으로 변경
