+++ src/App.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Routes, Route } from 'react-router-dom';
 import './App.css';
-import OnBoarding from './pages/onboarding/OnBoarding';
+import OnBoarding from './pages/onboarding/Onboarding';
 import Login from './pages/login/Login';
 import Home from './pages/home/Home';
 import FindCrew from './pages/findcrew/FindCrew';
