+++ src/pages/onboarding/OnBoardingStyle.css
